# api/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'College Attendance API'
  description: 'Versioned REST API for marking attendance, managing students/subjects, and generating detention & dashboard reports.'
  version: 1.0.0
servers:
  -
    url: /api/v1
    description: 'Default API server'
paths:
...
  /reports/detention:
    get:
      tags:
        - Detention
      summary: 'List detained students for a month (HOD or principal)'
      operationId: d7fca1780fdb52aa3d260aa0625c648a
      parameters:
        -
          name: month
          in: query
          description: 'Format Y-m, defaults to last month'
          required: false
          schema:
            type: string
            format: date
            example: 2026-05
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
            minimum: 1
        -
          name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: 'Paginated list of detained students'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    properties:
                      page:
                        type: integer
                      per_page:
                        type: integer
                      total:
                        type: integer
                      month:
                        type: string
                        example: 2026-05
                      threshold:
                        type: number
                        format: float
                        example: 75
                    type: object
                type: object
        403:
          description: 'Forbidden — hod or principal role required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        -
          bearerAuth: []
  /reports/detention/generate:
    post:
      tags:
        - Detention
      summary: 'Calculate detention for a month and optionally queue notification emails (principal only)'
      operationId: 2e598595635eec0e3552b9b1b3d0d563
      requestBody:
        required: false
        content:
          application/json:
            schema:
              properties:
                year:
                  type: integer
                  example: 2026
                  nullable: true
                month:
                  type: integer
                  example: 5
                  nullable: true
                send_emails:
                  type: boolean
                  example: false
                  nullable: true
              type: object
      responses:
        200:
          description: 'Detention report generated'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      message:
                        type: string
                      year:
                        type: integer
                      month:
                        type: integer
                      total_students:
                        type: integer
                      detained_count:
                        type: integer
                      emails_queued:
                        type: integer
                      emails_skipped:
                        type: integer
                      students:
                        type: array
                        items:
                          type: object
                    type: object
                type: object
        403:
          description: 'Forbidden — principal role required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        422:
          description: 'Validation failed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        -
          bearerAuth: []
  /hod/summary:
    post:
      tags:
        - HOD
      summary: 'Submit a daily attendance summary for a department/semester (HOD only)'
      operationId: a1ee16be711b6410b34bec274ac81fab
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - department
                - semester
                - year
                - date
                - total_students
                - present_count
              properties:
                department:
                  type: string
                  example: CSE
                semester:
                  type: integer
                  example: 4
                year:
                  type: integer
                  example: 2026
                date:
                  type: string
                  format: date
                  example: '2026-06-08'
                total_students:
                  type: integer
                  example: 60
                present_count:
                  type: integer
                  example: 54
              type: object
      responses:
        200:
          description: 'Summary submitted'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      message:
                        type: string
                      percentage:
                        type: number
                        format: float
                      summary:
                        type: object
                    type: object
                type: object
        403:
          description: 'Forbidden — hod role required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        422:
          description: 'Validation failed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        -
          bearerAuth: []
  '/reports/hod/{department}':
    get:
      tags:
        - HOD
      summary: 'List historical attendance summaries for a department (HOD or principal)'
      description: 'HODs may only view their own department; principals may view any department.'
      operationId: 7e17dcacbf0ce66b1b142489516af904
      parameters:
        -
          name: department
          in: path
          required: true
          schema:
            type: string
          example: CSE
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
            minimum: 1
        -
          name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: 'Paginated department attendance summaries'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    properties:
                      page:
                        type: integer
                      per_page:
                        type: integer
                      total:
                        type: integer
                    type: object
                type: object
        403:
          description: 'Forbidden — may only view own department'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
    ApiError:
      properties:
        success:
          type: boolean
          example: false
        error:
          properties:
            code:
              type: string
              example: VALIDATION_FAILED
            message:
              type: string
              example: 'The given data was invalid.'
            details:
              type: object
              nullable: true
          type: object
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Enter the JWT access token returned by /api/v1/auth/login'
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    bearerAuth: []